requirements.txt
asgiref~=3.3.1
certifi~=2020.12.5
chardet~=4.0.0
Django~=3.1.5
idna~=2.10
pycryptodome~=3.9.9
PyMySQL~=1.0.2
pytz~=2020.5
requests~=2.25.1
sqlparse~=0.4.1
urllib3~=1.26.2
beautifulsoup4~=4.9.3
lxml
readability-lxml